Python 2 and 3 work
